Franka-Teach/requirements.txt
pyzmq
scipy
opencv-python
blosc
hydra-core
pyrealsense2
h5py
pre-commit
